+++ src/CoercionMode.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Duon\Sire;
+
+/** @api */
+enum CoercionMode
+{
+	case Coerce;
+	case Strict;
+}