+++ src/Contract/Coercer.php
@@ -4,10 +4,15 @@ declare(strict_types=1);
 
 namespace Duon\Sire\Contract;
 
+use Duon\Sire\CoercionMode;
+
 /** @api */
 interface Coercer
 {
 	public string $message { get; }
 
-	public function coerce(mixed $pristine): Coercion;
+	public function coerce(
+		mixed $pristine,
+		CoercionMode $mode = CoercionMode::Coerce,
+	): Coercion;
 }